+++ src/main/nextflow/co2footprint/CO2FootprintCalculator.groovy
@@ -109,8 +109,6 @@ class CO2FootprintCalculator {
             throw new MissingValueException(message)
         }
 
-        final BigDecimal powerdrawMem  = config.powerdrawMem // [W per GB]
-
         /* ===== Data Center Effectiveness and Carbon Intensity ===== */
 
          // PUE: power usage effectiveness of datacenter [ratio] (>= 1.0)
@@ -126,14 +124,18 @@ class CO2FootprintCalculator {
         /* ===== Energy & Emission Calculation ===== */
 
         // Energy consumption [kWh]
-        BigDecimal rawEnergyProcessor
-        if (config.cpuPowerModel) {
-            rawEnergyProcessor = runtime_h * numberOfCores * config.cpuPowerModel(coreUsage) * 0.001
-        }
-        else {
-            rawEnergyProcessor = runtime_h * numberOfCores * powerdrawPerCore * coreUsage * 0.001
-        }
-        BigDecimal rawEnergyMemory = runtime_h * memory * powerdrawMem * 0.001
+        String cpuEnergyFunction = config.cpuEnergyFunction ?: 'runtime_h * numberOfCores * powerdrawPerCore * coreUsage'
+        BigDecimal rawEnergyProcessor = evaluateStringCalculation(
+                cpuEnergyFunction,
+                [runtime_h: runtime_h, numberOfCores: numberOfCores, powerdrawPerCore: powerdrawPerCore, coreUsage: coreUsage]
+        ) * 0.001
+
+        String memoryEnergyFunction = config.memoryEnergyFunction ?: 'runtime_h * memory * 0.3725'
+        BigDecimal rawEnergyMemory = evaluateStringCalculation(
+                memoryEnergyFunction,
+                [runtime_h: runtime_h, memory: memory]
+        ) * 0.001
+                
         BigDecimal energy = pue * (rawEnergyProcessor + rawEnergyMemory)
 
         // Resulting CO₂ emissions
@@ -154,6 +156,8 @@ class CO2FootprintCalculator {
             config.ignoreCpuModel ? 'Custom value' : cpuModel,
             rawEnergyProcessor,
             rawEnergyMemory,
+            cpuEnergyFunction,
+            memoryEnergyFunction
         )
     }
 
@@ -206,5 +210,16 @@ class CO2FootprintCalculator {
             )
         }
         return value != null ? value : defaultValue
+    }
+
+    /**
+     * Evaluate a string to Groovy code and execute it with the given context parameters.
+     * 
+     * @param functionString A function in String form
+     * @param context parameters that are applied to the given function
+     * @return The result as a BigDecimal number
+     */
+    private static BigDecimal evaluateStringCalculation(String functionString, Map<String, Object> context) {
+        return new GroovyShell(new Binding(context)).evaluate(functionString) as BigDecimal
     }
 }