+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ## Features:
 - Transformation of data file to provenance file with schema.org / bioschemas.org type annotation in JSON-LD data format
+- Added functionality to apply fully custom memory and CPU power functions
 
 # 1.2.1
 ## Bug Fixes: